address review comments: - typos and such - reorganize headers - initialize debug variables in a specific init.cpp file - refactor conditions in peek/poke alignment handling - factorize common code in ProcessCommand and SendError - move generic binary handling functions from debug interface to common/libs/binary_utils make commands registry smaller
