+++ software/controller/src/main.cpp
@@ -16,9 +16,9 @@ limitations under the License.
 #include "actuators.h"
 #include "comms.h"
 #include "controller.h"
-#include "debug.h"
 #include "eeprom.h"
 #include "hal.h"
+#include "interface.h"
 #include "network_protocol.pb.h"
 #include "nvparams.h"
 #include "sensors.h"
